pyproject.toml
[tool.poetry]
name = "my_scraper_project"
version = "0.1.0"
description = "A web scraping project"
authors = ["Your Name <youremail@example.com>"]
packages = [
    { include = "src" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.0.0"
beautifulsoup4 = "^4.9.3"
pandas = "^2.2.2"
pymysql = "^1.0.2"
hydra-core = "^1.1.0"
omegaconf = "^2.1.1"
webdriver-manager = "^4.0.2"
loguru = "^0.7.2"

mysql-connector-python = "^8.0"
black = "^24.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.pytest.ini_options]
testpaths = ["tests"]